Create _strip-unit.scss
Add a strip unit helper that will be used for the calculation of the relative units. With this helper, we can pass numbers with units (e.g `20px`) or without, when calculating relative values (those will have to be changed of course). That way we can reuse things from `manifest.json`.