Return a blank example as an intermediary fix for example crashing
The example currently crashes as line 9 returns null in the example renderer. Until further feedback in PR comments, I've put this in to prevent an ugly error when hovering.

A dummy example could be set, but we should talk about how useful that would be. I've left a comment regarding this on @iruzevic's review.