Remove console logs
In stories the logs are replaced by info, and in one helper it was removed. It outputted tons of info in the console. This one slipped through our fingers. I noticed that the linters are not picking them up, and they should. So maybe we need to fix them to catch these cases.