Fix Yoast analysis
Improves the regex for capturing content from the attributes. Added a way to include image URLs so those are also included in the analysis. We should see how to add image alt, since Yoast probably analyses this as well. I also replaced the unicode characters with their non-unicode counterparts, since that prevented Yoast from detecting outbound and internal links. This was tested on Yoast Pro version. The last thing that probably should get improved is ordering the content, because I think this just parses blocks in some weird order. And headings are not wrapped in correct h tags which also skewes the analysis.