Set iconUse in button component to false by default
Buttons that have iconUse set to true, but an icon isn't selected, gain
extra visual padding on the left; i.e. the icon takes up space, although
it isn't visible. This is a visual issue, and instead of setting a
default icon, I've decided to set iconUse to false by default.

The reasoning behind setting iconUse to false instead of setting a
default icon is that I wasn't sure which icon could sensibly serve
as a default, and also because I believe this is more DX-friendly.
I'm fine with a different approach as well, this is just the simplest
fix I could think of.